Update 22/05/2022 20:32
Images now load off a URL, home screen finished (Search Bar WIP).
Working on loading forum threads, when a topic is clicked the id of the topic is saved and can be used to load the data with Realtime Database, also loads the title, though it should also be loaded via Realtime Database.